Add support for ProfileCredentialsProvider
This commit resolves spring-cloud/spring-cloud-aws#52 by adding support to specify a ProfileCrednetialsProvider. This can be done with an XML-Element <context:profile-credentials profileName="test" />, the Java Annotation EnableContextCredentials and also the auto configuration with the attributes cloud.aws.credentials.profileName and cloud.aws.credentials.profilePath

It is possible to configure the profile name using the default file (which is in the user home), or to specify the profile file itself.