Improve warning when SQS queue resolving fails. This can also fail when no access is granted to perform sqs:GetQueueUrl.
Before this commit the following message is provided "Ignoring queue with name 'queue name' as it does not exist.". This can be misleading when the queue does exist, but there is no permission provided
to perform "sqs:GetQueueUrl". The AWS SDK's QueueDoesNotExistException message is different when the potentially there is no access to perform GetQueueUrl: "The specified queue does not exist or you do not have access to it.". To make it clear exactly what is failing (GetQueueUrl) a custom message is provided.